server/requirements.txt
appnope==0.1.0
astroid==1.4.4
backports-abc==0.4
colorama==0.3.6
decorator==4.0.9
gnureadline==6.3.3
ipython==4.1.1
ipython-genutils==0.1.0
lazy-object-proxy==1.2.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
pylint==1.5.4
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
tornado==4.3
traitlets==4.1.0
websocket-client==0.35.0
wrapt==1.10.6
